Further refactoring, command upgrades
Various refactoring for code clarity.

replaceblocks and countblocks now support targeting any singular entity. The direct coordinate variant now uses BlockPos as the argument type, allowing for the use of relative positioning (via "~") in addition to direct coordinates.

The countblocks command now holds the breakdown of the block types. This breakdown can be accessed in pages via the countblockslist command.